Add system property for skipping the Mock unit tests.
git-svn-id: https://svn.apache.org/repos/asf/jakarta/tapestry/trunk@244151 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/src/test/org/apache/tapestry/junit/mock/TestMocks.java
@@ -69,10 +69,17 @@ public class TestMocks extends TapestryTestCase
     {
         TestSuite suite = new TestSuite("Mock Unit Test Suite");
 
-        addScripts(suite);
+        if (Boolean.getBoolean("skip-mock-tests"))
+        {
+            System.out.println("*** Skipping Mock Unit Test Suite");
+        }
+        else
+        {
+            addScripts(suite);
 
-        // Handy place to perform one-time
-        deleteDir(getBaseDirectory() + "/target/.private");
+            // Handy place to perform one-time
+            deleteDir(getBaseDirectory() + "/target/.private");
+        }
 
         return suite;
     }
